:truck: chore(api): move student certification to `src`
