+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -2,25 +2,6 @@ import * as divisionSerializer from '../../../src/prescription/campaign/infrastr
 import * as certificationCenterMembershipSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
 import { usecases as teamUsecases } from '../../../src/team/domain/usecases/index.js';
 import { usecases } from '../../domain/usecases/index.js';
-import * as studentCertificationSerializer from '../../infrastructure/serializers/jsonapi/student-certification-serializer.js';
-
-const getStudents = async function (request) {
-  const certificationCenterId = request.params.certificationCenterId;
-  const sessionId = request.params.sessionId;
-
-  const { filter, page } = request.query;
-  if (filter.divisions && !Array.isArray(filter.divisions)) {
-    filter.divisions = [filter.divisions];
-  }
-
-  const { data, pagination } = await usecases.findStudentsForEnrolment({
-    certificationCenterId,
-    sessionId,
-    page,
-    filter,
-  });
-  return studentCertificationSerializer.serialize(data, pagination);
-};
 
 const getDivisions = async function (request) {
   const certificationCenterId = request.params.certificationCenterId;
@@ -77,7 +58,6 @@ const certificationCenterController = {
   createCertificationCenterMembershipByEmail,
   findCertificationCenterMembershipsByCertificationCenter,
   getDivisions,
-  getStudents,
   updateReferer,
 };
