+++ api/src/certification/enrolment/application/certification-center-route.js
@@ -1,0 +1,46 @@
+import Joi from 'joi';
+
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
+import { certificationCenterController } from './certification-center-controller.js';
+
+const register = async function (server) {
+  server.route([
+    {
+      method: 'GET',
+      path: '/api/certification-centers/{certificationCenterId}/sessions/{sessionId}/students',
+      config: {
+        validate: {
+          params: Joi.object({
+            certificationCenterId: identifiersType.certificationCenterId,
+            sessionId: identifiersType.sessionId,
+          }),
+          query: Joi.object({
+            filter: Joi.object({
+              divisions: Joi.array().items(Joi.string()),
+            }).default({}),
+            page: {
+              number: Joi.number().integer(),
+              size: Joi.number().integer(),
+            },
+          }),
+        },
+        pre: [
+          {
+            method: securityPreHandlers.checkUserIsMemberOfCertificationCenter,
+            assign: 'isMemberOfCertificationCenter',
+          },
+        ],
+        handler: certificationCenterController.getStudents,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Récupération d'une liste d'élèves SCO à partir d'un identifiant de centre de certification",
+        ],
+        tags: ['api', 'certification-center', 'students', 'session'],
+      },
+    },
+  ]);
+};
+
+const name = 'certification-centers-session-students-api';
+export { name, register };
