+++ api/src/certification/enrolment/domain/usecases/find-students-for-enrolment.js
@@ -1,5 +1,5 @@
-import { NotFoundError } from '../../../src/shared/domain/errors.js';
-import { StudentForEnrolment } from '../../../src/shared/domain/read-models/StudentForEnrolment.js';
+import { NotFoundError } from '../../../../shared/domain/errors.js';
+import { StudentForEnrolment } from '../../../../shared/domain/read-models/StudentForEnrolment.js';
 
 const findStudentsForEnrolment = async function ({
   certificationCenterId,
