+++ api/src/certification/enrolment/infrastructure/serializers/student-certification-serializer.js
