+++ api/src/certification/enrolment/routes.js
@@ -1,5 +1,6 @@
 import * as attendanceSheet from './application/attendance-sheet-route.js';
 import * as certificationCandidate from './application/certification-candidate-route.js';
+import * as certificationCenter from './application/certification-center-route.js';
 import * as country from './application/country-route.js';
 import * as enrolment from './application/enrolment-route.js';
 import * as sessionMassImport from './application/session-mass-import-route.js';
@@ -14,6 +15,7 @@ const certificationEnrolmentRoutes = [
   enrolment,
   session,
   sessionMassImport,
+  certificationCenter,
   subscription,
   user,
 ];
