+++ api/tests/acceptance/application/certification-centers/certification-center-controller_test.js
@@ -49,104 +49,6 @@ describe('Acceptance | API | Certification Center', function () {
     });
   });
 
-  describe('GET /api/certification-centers/{certificationCenterId}/sessions/{sessionId}/students', function () {
-    let request;
-    const externalId = 'XXXX';
-
-    function _buildOrganizationLearnersWithConnectedUserRequest(user, certificationCenter, session) {
-      return {
-        method: 'GET',
-        url:
-          '/api/certification-centers/' +
-          certificationCenter.id +
-          '/sessions/' +
-          session.id +
-          '/students?page[size]=10&page[number]=1',
-        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
-      };
-    }
-    function _buildOrganizationLearnersNotConnectedUserRequest(certificationCenter, session) {
-      return {
-        method: 'GET',
-        url:
-          '/api/certification-centers/' +
-          certificationCenter.id +
-          '/sessions/' +
-          session.id +
-          '/students?page[size]=10&page[number]=1',
-      };
-    }
-
-    context('when user is connected', function () {
-      it('should return 200 HTTP status', async function () {
-        // given
-        const { certificationCenter, user } = _buildUserWithCertificationCenterMemberShip(externalId);
-        const organization = databaseBuilder.factory.buildOrganization({ externalId, type: 'SCO' });
-        const session = databaseBuilder.factory.buildSession({ certificationCenterId: certificationCenter.id });
-        _buildOrganizationLearners(organization, { firstName: 'Laura', lastName: 'certifForEver', division: '2ndB' });
-        await databaseBuilder.commit();
-
-        const request = _buildOrganizationLearnersWithConnectedUserRequest(user, certificationCenter, session);
-
-        // when
-        const response = await server.inject(request);
-
-        // then
-        expect(response.statusCode).to.equal(200);
-      });
-
-      it('should return the organization learners asked', async function () {
-        // given
-        const { certificationCenter, user } = _buildUserWithCertificationCenterMemberShip(externalId);
-        const organization = databaseBuilder.factory.buildOrganization({ type: 'SCO', externalId });
-        const session = databaseBuilder.factory.buildSession({ certificationCenterId: certificationCenter.id });
-        _buildOrganizationLearners(
-          organization,
-          { id: 1, division: '2ndB', firstName: 'Laura', lastName: 'Certif4Ever' },
-          { id: 2, division: '2ndA', firstName: 'Laura', lastName: 'Booooo' },
-          { id: 3, division: '2ndA', firstName: 'Laura', lastName: 'aaaaa' },
-          { id: 4, division: '2ndA', firstName: 'Bart', lastName: 'Coucou' },
-          { id: 5, division: '2ndA', firstName: 'Arthur', lastName: 'Coucou' },
-        );
-        await databaseBuilder.commit();
-
-        const request = _buildOrganizationLearnersWithConnectedUserRequest(user, certificationCenter, session);
-
-        // when
-        const response = await server.inject(request);
-
-        // then
-        expect(_.map(response.result.data, 'id')).to.deep.equal(['3', '2', '5', '4', '1']);
-      });
-    });
-
-    context('when user is not connected', function () {
-      it('should return 401 HTTP status code if user is not authenticated', async function () {
-        // given
-        _buildUserWithCertificationCenterMemberShip(externalId);
-        databaseBuilder.factory.buildOrganization({ externalId });
-        const certificationCenterWhereUserDoesNotHaveAccess = databaseBuilder.factory.buildCertificationCenter({
-          externalId,
-        });
-        const session = databaseBuilder.factory.buildSession({
-          certificationCenterId: certificationCenterWhereUserDoesNotHaveAccess.id,
-        });
-        await databaseBuilder.commit();
-
-        request = _buildOrganizationLearnersNotConnectedUserRequest(
-          certificationCenterWhereUserDoesNotHaveAccess,
-          session,
-        );
-
-        // when
-        const response = await server.inject(request);
-
-        // then
-        expect(response.statusCode).to.equal(401);
-      });
-    });
-  });
-
   describe('GET /api/admin/certification-centers/{id}/certification-center-memberships', function () {
     context('when certification center membership is linked to the certification center', function () {
       it('should return 200 HTTP status', async function () {
