+++ api/tests/certification/enrolment/unit/domain/usecases/find-students-for-enrolment_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 
-import { findStudentsForEnrolment } from '../../../../lib/domain/usecases/find-students-for-enrolment.js';
-import { NotFoundError } from '../../../../src/shared/domain/errors.js';
-import { StudentForEnrolment } from '../../../../src/shared/domain/read-models/StudentForEnrolment.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { findStudentsForEnrolment } from '../../../../../../src/certification/enrolment/domain/usecases/find-students-for-enrolment.js';
+import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
+import { StudentForEnrolment } from '../../../../../../src/shared/domain/read-models/StudentForEnrolment.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | find-students-for-enrolment', function () {
   const certificationCenterId = 1;
