+++ api/tests/certification/enrolment/unit/infrastructure/serializers/student-certification-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/student-certification-serializer.js';
-import { StudentForEnrolment } from '../../../../../src/shared/domain/read-models/StudentForEnrolment.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/certification/enrolment/infrastructure/serializers/student-certification-serializer.js';
+import { StudentForEnrolment } from '../../../../../../src/shared/domain/read-models/StudentForEnrolment.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | student-certification-serializer', function () {
   describe('#serialize', function () {
