+++ api/tests/shared/integration/infrastructure/repositories/organization-learner-repository_test.js
@@ -1,12 +1,12 @@
 import dayjs from 'dayjs';
 import _ from 'lodash';
 
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
-import * as organizationLearnerRepository from '../../../../lib/infrastructure/repositories/organization-learner-repository.js';
-import { ORGANIZATION_FEATURE } from '../../../../src/shared/domain/constants.js';
-import { OrganizationLearnerNotFound, UserNotFoundError } from '../../../../src/shared/domain/errors.js';
-import { OrganizationLearner } from '../../../../src/shared/domain/models/index.js';
-import { catchErr, databaseBuilder, domainBuilder, expect, knex, sinon } from '../../../test-helper.js';
+import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import { ORGANIZATION_FEATURE } from '../../../../../src/shared/domain/constants.js';
+import { OrganizationLearnerNotFound, UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
+import { OrganizationLearner } from '../../../../../src/shared/domain/models/index.js';
+import * as organizationLearnerRepository from '../../../../../src/shared/infrastructure/repositories/organization-learner-repository.js';
+import { catchErr, databaseBuilder, domainBuilder, expect, knex, sinon } from '../../../../test-helper.js';
 
 describe('Integration ¨| Infrastructure | Repository | organization-learner-repository', function () {
   describe('#findByIds', function () {
